test and document createPureReaction
