+++ packages/memo/README.md
@@ -11,6 +11,7 @@ Collection of custom `createMemo` primitives. They extend it's functionality whi
 - [`createAsyncMemo`](#createAsyncMemo) - Memo that allows for asynchronous calculations.
 - [`createDebouncedMemo`](#createDebouncedMemo) - Memo which returned signal is debounced.
 - [`createThrottledMemo`](#createThrottledMemo) - Memo which returned signal is throttled.
+- [`createPureReaction`](#createPureReaction) - A `createReaction` that runs before render _(non-batching)_.
 
 ## Installation
 
@@ -134,6 +135,32 @@ const double = createThrottledMemo(prev => count() * 2, 200, { value: 0 });
 ```
 
 Note: the callback is run initially to kickoff tracking and set the signal's value.
+
+## `createPureReaction`
+
+Solid's [`createReaction`](#https://www.solidjs.com/docs/latest/api#createreaction) that is based on pure computation _(runs before render, and is non-batching)_
+
+### How to use it
+
+It's usage exactly matches the original. The only difference is in when the callback is being executed, the normal createReaction runs it after render, similar to how effects work, while the createPureReaction is more like createComputed.
+
+```ts
+import { createPureReaction } from "@solid-primitives/memo"
+
+const [count, setCount] = createSignal(0);
+const track = createPureReaction(() => {...});
+track(count);
+setCount(1); // triggers callback
+
+// sources need to be re-tracked every time
+setCount(2); // doesn't trigger callback
+```
+
+### Definition
+
+```ts
+function createPureReaction(onInvalidate: Fn, options?: EffectOptions): (tracking: Fn) => void;
+```
 
 ## Changelog
